Exclude "waiting on response" issues from unanswered questions.
This is useful for questions where the most recent comment was from the user saying something like "okay, I will try that and get back to you".  We may want to remove it from the list and also start the timer for the bot to come along and auto-close it.